+++ gadd/findNumber/NumberFinderNoThreads.java
@@ -1,21 +1,13 @@
-package gadd.FindNumber;
+package gadd.findNumber;
 
-import java.util.Arrays;
 import java.util.HashSet;
-import java.util.Set;
 
 
 public class NumberFinderNoThreads {
-//    private long start = 1234567890L, stop = 9876543210L;
-
-    public NumberFinderNoThreads() {
-//        go(start, stop);
-    }
 
     protected boolean checkDifferentDigits(long numberToCheck) {
         String numberString = Long.toString(numberToCheck); // convert number to string
         char[] numberArray = numberString.toCharArray(); // convert string to array
-//            Set numberSet = new HashSet<>(Arrays.asList(numberArray)); // convert array to set !!! IT IS NOT WORKING!!! (first element - is set)
         HashSet numberSet = new HashSet();
         for (char ch : numberArray) {
             numberSet.add(ch);
@@ -34,15 +26,15 @@ public class NumberFinderNoThreads {
     }
 
 
-    protected void go(long startNumber, long stopNumber) {
+    protected long go(long startNumber, long stopNumber) {
         for (long i = startNumber; i <= stopNumber; i += 10) {
             if (checkDifferentDigits(i)) {
                 if (checkRule(i)) {
-                    System.out.println(i);
+                    return (i);
                 }
-
             }
         }
+        return 0;
     }
