+++ gadd/findNumber/Run.java
@@ -1,4 +1,4 @@
-package gadd.FindNumber;
+package gadd.findNumber;
 
 public class Run {
     public static void main(String[] args) {
@@ -6,13 +6,18 @@ public class Run {
 
         System.out.println("Without threads starts");
         startTime = System.currentTimeMillis();
-        new NumberFinderNoThreads().go(1234567890L, 9876543210L);
+
+        long result = new NumberFinderNoThreads().go(1234567890L, 9876543210L);
+        if (result != 0) System.out.println(result);
+        else System.out.println("Nothing found!");
+
         stopTime = System.currentTimeMillis();
         System.out.println("Without threads ends in " + (stopTime - startTime) + "ms");
 
         int numberOfThreads = 5;
         startTime = System.currentTimeMillis();
         System.out.println("\nWith " + numberOfThreads + " threads starts in " + startTime);
+
         long start = 123456789L, stop = 987654321L;
         long step = (stop - start) / numberOfThreads;
         for (int i = 0; i < numberOfThreads; i++) {
